test: flatten vm submodule under imgtestlib
Flatten the vm module (previously vmtest) under imgtestlib.  All
functions are available under imgtestlib.vm now.
Move all vm tests (previously vmtest/test) to test_imgtestlib.
Update import paths in the rest of the imgtestlib module to import the
vm submodule.
